Enable dual callback/promise mode on model methods
When the method arguments do not provide any callback argument,
then return a promise that will be resolved/rejected with the operation
outcome.

Also correctly handle the case when the last argument is an options
object enabling the mock mode - in such case the result of the operation
is returned directly.